Crate bitbazaar

bitbazaar - An assortment of publicly available cross-language utilities useful to my projects.

Modules

chrono

Chrono utilities

cli

Command line interface utilities.

command

OS Command related utilities. Much lighter weight than cli module (which will probably be deprecated as badly written and not maintainable). Not applicable to wasm

cookies

Setting/getting cookies in wasm or ssr.

crypto

Hashing & encryption utilities. Most inside is feature gated.

errors

Error handling utilities.

file

File related utilities. Therefore disabled on wasm.

log

Logging utilities

misc

Completely miscellaneous utilities

redis

Redis utilities

test

Useful and reusable testing utilities. Should only be enabled as a feature in dev-dependencies.

threads

Concurrency/parallelism utilities

timing

Timing utilities

Macros

anyerr

A macro for building Report<AnyErr> objects with string information easily.

assert_td_in_range

Asserts that the given chrono::TimeDelta is in the given range.

err

A macro for building Report<ArbitraryErrorStackErr> objects with string context easily.

panic_on_err

When working in a function that cannot return a result, use this to auto panic with the formatted error if something goes wrong.

panic_on_err_async

When working in a function that cannot return a result, use this to auto panic with the formatted error if something goes wrong.

retry_flexi

Internal version of Retry::call that:

timeit

Time a block of code and log to the global time recorder.

warn_user_once

Warn a user once, with uniqueness determined by the calling location itself.

warn_user_once_by_id

Warn a user once, with uniqueness determined by the given ID.