Trait bitbazaar::command::CmdSpawnExt

pub trait CmdSpawnExt {
    type Spawner<'a>
       where Self: 'a;

    // Required method
    fn spawn_builder(&mut self) -> Self::Spawner<'_>;
}

Extension trait for getting the spawn extension builder from both std and tokio commands.

Required Associated Types

type Spawner<'a> where Self: 'a

The type of the spawn extension builder.

Required Methods

fn spawn_builder(&mut self) -> Self::Spawner<'_>

Get a spawn extension builder that provides the ability to:

Listen to stdout and/or stderr line by line in real-time via callbacks.

Object Safety

This trait is not object safe.