Module bitbazaar::cookies

Setting/getting cookies in wasm or ssr.

Structs

CookieOptions

Cookies options (see https://developer.mozilla.org/en-US/docs/Web/API/Document/cookie).

Enums

SameSite

SameSite value for CookieOptions.

Functions

delete_cookie

Delete a cookie if it exists. Cookies might not delete if path or domain are different, if not deleting pass the same options.

get_cookie

Get and automatically decode a cookie into a deserializable type. If the cookie isn’t found, or if it fails to deserialize, returns None. When it fails to deserialize, an error will be recorded.

get_cookie_raw

Get the raw value of a cookie. If the cookie isn’t found, returns None.

set_cookie

Set a new cookie with the given name and serializable value. If serialization fails, an error will be recorded.

set_cookie_raw

Set a new cookie with the given name and raw value.