Function bitbazaar::crypto::md5::md5_raw

pub fn md5_raw<B: AsRef<[u8]>>(
    inputs: impl IntoIterator<Item = B>,
) -> impl LowerHex + AsRef<[u8]> + 'static

md5 hash function, lower level than [md5] to allow multiple inputs and byte output.

NOTE: this is an insecure hash function from modern standards. Only use where you have to for backwards compatibility.

Arguments

inputs: The input data.