Trait bitbazaar::errors::BitbazaarResultExt

pub trait BitbazaarResultExt {
    type Context: Context;
    type Ok;

    // Required method
    fn loc(self) -> Result<Self::Ok, Report<Self::Context>>;
}

Further extensions on top of error_stack::ResultExt.

Required Associated Types

type Context: Context

The [Context] type of the Result.

type Ok

Type of the Ok value in the Result

Required Methods

fn loc(self) -> Result<Self::Ok, Report<Self::Context>>

Attach the current location to the stacktrace of the Report inside the Result if it’s in error state. This is useful in 2 cases:

Async functions, therefore can’t use #[track_caller] yet, once stable those uses can be removed. See https://github.com/rust-lang/rust/issues/110011.

When the same error is propagating through a middleman fn, and we really want to know the middleman specific location.