Module bitbazaar::log

Logging utilities

Modules

otlp

Opentelemetry types that might be needed downstream. The aim is to avoid the user having to depend on opentelemetry crates directly.

Structs

GlobalLog

The global logger/tracer for stdout, file and full open telemetry. Works with the tracing crates (info!, debug!, warn!, error!) and span funcs and decorators.

GlobalLogBuilder

The global log builder. See the GlobalLog struct for more information.

Functions

flush

Force through logs, traces and metrics, useful in e.g. testing.

global_meter

Returns the default [opentelemetry::metrics::Meter] for the app, labelled “default”.

init_system_and_process_metrics

Automatically record system metrics: SYSTEM WIDE:

meter

Returns a new [opentelemetry::metrics::Meter] with the provided name and default configuration.

record_exception

Record an exception to the currently active span. Matches oltp spec so it shows up correctly as an exception in observers https://opentelemetry.io/docs/specs/semconv/exceptions/exceptions-spans/

set_response_headers_from_ctx

Set the response headers from the current span context. So downstream services can continue the current trace.

set_span_parent_from_http_headers

Connect this program’s span to the trace that is represented by the provided HTTP headers. E.g. connect an axum handler’s trace/span to the nginx trace/span.

shutdown

Shutdown the logger, traces and metrics, should be called when the program is about to exit.