Function bitbazaar::log::record_exception

pub fn record_exception(
    message: impl Into<String>,
    stacktrace: impl Into<String>
)

Record an exception to the currently active span. Matches oltp spec so it shows up correctly as an exception in observers https://opentelemetry.io/docs/specs/semconv/exceptions/exceptions-spans/

Arguments:

message: Information about the exception e.g. Internal Error leading to 500 http response.

stacktrace: All of the location information for the exception, (maybe also the exception itself if e.g. from Report<T>).