Function bitbazaar::logging::create_subscriber

pub fn create_subscriber(
    layers: Vec<SubLayer>
) -> Result<CreatedSubscriber, AnyErr>

Simple interface to setup a sub and output to a given target. Returns the sub, must run sub.apply()? To actually enable it as the global sub, this can only be done once.

Example sub with an stdout logger:

use bitbazaar::logging::{SubLayer, create_subscriber, SubLayerFilter};
use tracing_subscriber::prelude::*;
use tracing::Level;

let sub = create_subscriber(vec![SubLayer {
    filter: SubLayerFilter::Above(Level::INFO), // Only log info and above
    ..Default::default()
}]).unwrap();
sub.into_global(); // Register it as the global sub, this can only be done once

Example sub with a file logger:

use bitbazaar::logging::{SubLayer, SubLayerVariant, create_subscriber, SubLayerFilter};
use tracing_subscriber::prelude::*;
use tracing::Level;

let temp_dir = tempfile::tempdir().unwrap();
let sub = create_subscriber(vec![SubLayer {
    filter: SubLayerFilter::Above(Level::INFO), // Only log info and above
    variant: SubLayerVariant::File {
        file_prefix: "my_program_".into(),
        dir: temp_dir.path().to_path_buf(),
    },
    ..Default::default()
}]).unwrap();
sub.into_global(); // Register it as the global sub, this can only be done once

Example sub with a custom writer:

use bitbazaar::logging::{SubCustomWriter, create_subscriber, SubLayer, SubLayerVariant, SubLayerFilter};
use tracing_subscriber::prelude::*;
use tracing::Level;

let sub = create_subscriber(vec![SubLayer {
   filter: SubLayerFilter::Above(Level::INFO), // Only log info and above
   variant: SubLayerVariant::Custom {
       writer: SubCustomWriter {
           write: |log| {
               println!("My custom log: {}", String::from_utf8_lossy(log));
           },
       },
       include_color: false,
   },
   ..Default::default()
}]).unwrap();
sub.into_global(); // Register it as the global sub, this can only be done once

Example sub with open telemetry: E.g. jaeger running locally on localhost:4317 (docker pull jaegertracing/all-in-one:latest && docker run -d --name jaeger -e COLLECTOR_OTLP_ENABLED=true -p 16686:16686 -p 4317:4317 -p 4318:4318 jaegertracing/all-in-one:latest)

use bitbazaar::logging::{SubCustomWriter, create_subscriber, SubLayer, SubLayerVariant, SubLayerFilter};
use tracing_subscriber::prelude::*;
use tracing::Level;

let sub = create_subscriber(vec![SubLayer {
   filter: SubLayerFilter::Above(Level::INFO), // Only log info and above
   variant: SubLayerVariant::OpenTelemetry {
       endpoint: "http://localhost:4317".into(),
       headers: vec![], // Additional headers to send to the provider
   },
   ..Default::default()
}]).unwrap();
sub.into_global(); // Register it as the global logger, this can only be done once