Struct bitbazaar::logging::SubCustomWriter

pub struct SubCustomWriter {
    pub write: fn(_: &[u8]),
}

A custom writer for logging, to something other than the provided variants.

See the create_subscriber fn for examples.

Fields
write: fn(_: &[u8])

The fn to handle writing, passed the raw byte string. If needing a string, can do:

let log = String::from_utf8_lossy(log);