Module bitbazaar::misc

Completely miscellaneous utilities

Modules

bytes

Byte manipulation utilities, e.g. transfer speed.

platform

Platform utilities, e.g. OS type, cpu arch, in_ci.

Macros

retry_flexi

Internal version of Retry::call that:

Structs

FlexiLogBridge

Flexi logging bridge to wrap arbitrary flexi logging implementers, doesn’t require any generics.

FlexiLogBridgeWriter

Flexi logging bridge writer, doesn’t require any generics.

LocalPeriodicUpdater

A periodic updater. Run a callback at a specified time interval. Synchronous. Requires polling. Useful in long running loops to do something at a specified time interval. PeriodicUpdater::maybe_update should be called during each loop.

Looper

Standardizes pattern for internalised loop that allows mutable callbacks and persistent state. The state is data that’s persisted between iterations, and eventually passed out, this should be customisable by the caller.

PeriodicUpdater

A periodic updater. Run a callback at a specified time interval. Synchronous. Requires polling. Useful in long running loops to do something at a specified time interval. PeriodicUpdater::maybe_update should be called during each loop.

PeriodicUpdaterState

State passed into the callback of PeriodicUpdater.

Refreshable

A data wrapper that automatically updates the data given out when deemed stale. The data is set to refresh at a certain interval (triggered on access), or can be forcefully refreshed.

RefreshableGuard

A temporary storage of the pointer.

Retry

The retry builder.

RetryInfo

Information about the last retry attempt.

TarballFile

A decompressed file from a tarball wrapped in a simpler interface. Implements std::io::Read for easy lazy reading.

Enums

Backoff

UNSTABLE: ONLY PUBLIC FOR MACRO USE.

FlexiLogPhase

Useful for phase tracking, depending on the logger.

RetryUntil

UNSTABLE: ONLY PUBLIC FOR MACRO USE.

Traits

BinarySearchable

A trait allowing arbitrary types to be binary searchable.

FlexiLog

An interface that can be used to track arbitrary logging “flows”.

FlexiLogFromRedis

Lots of loggers are backed by redis. A higher-order trait to produce a logger once redis is available.

FlexiLogWriter

The actual object used to apply logs. This is useful when updates are async, making synchronous batching easy.

SerdeMigratable

A trait to help with migrating data structures when they change.

Functions

binary_search_exact

Binary search for the desired item in the array. Returns None when not exactly found.

binary_search_exact_with_siblings

Binary search for the desired item in the array, but also passes the prior and next items into the comparison function, if they exist. Returns None when not exactly found.

binary_search_soft

Binary search for the desired item in the array. Only returns None when the array is empty.

global_lock_host_async

APPLIES TO ALL PROCESSES ON HOST MACHINE. If only needing for the current process, use global_lock_process_async instead.

global_lock_host_sync

APPLIES TO ALL PROCESSES ON HOST MACHINE. If only needing for the current process, use global_lock_process_sync instead.

global_lock_process_async

APPLIES TO ONLY THE CURRENT PROCESS. If wanting to apply to all processes on the host machine, use [global_lock_host] instead.

global_lock_process_sync

APPLIES TO ONLY THE CURRENT PROCESS. If wanting to apply to all processes on the host machine, use [global_lock_host] instead.

is_tcp_port_listening

Check if a port is listening for a given ipv4 address and port.

main_wrapper

Wrap your main function in this to auto handle common program needs.

random_bytes_insecure_speedy

Generate random bytes using the system’s random number generator. Insecure, but very fast, useful for testing random binary data.

random_u64_rolling

Fastest possible unique u64. Just increments an atomic, making sure to roll at max.

setup_once

Not available on wasm, requires global_lock_host_async which is filesystem based. Setup something once for a given version on the current host. E.g. installing a package, downloading a file, etc.

setup_once_clear

Clear all data for a given id.

setup_once_storage_path

Get the outer path all data is stored in.

sleep_compat

Sleep for a duration, compatible with both WASM and native targets.

tarball_compress

Compress a tarball (.tar.gz).

tarball_decompress

Decompress a tarball (.tar.gz).

with_timeout

Run the future with the given timeout, returning the result of the closure and cancelling the future if it takes too long.