Function bitbazaar::misc::binary_search_exact_with_siblings

pub fn binary_search_exact_with_siblings<Item>(
    arr_like: &impl BinarySearchable<Item = Item>,
    comparer: impl Fn(Option<&Item>, &Item, Option<&Item>) -> Ordering,
) -> Option<usize>

Binary search for the desired item in the array, but also passes the prior and next items into the comparison function, if they exist. Returns None when not exactly found.

Arguments:

arr_like: The array-like structure to search.

comparer: Fn(Option<&Item>, &Item, Option<&Item>) (prior, target, next) that should return $target.cmp($item_val). (i.e. “where to go from here”)

Returns:

Some(index) the index of the item in the array.

None if the item is not found or array empty.