Trait bitbazaar::misc::FlexiLog

pub trait FlexiLog: Debug + Sync + Send {
    type Writer: FlexiLogWriter;

    // Required methods
    fn batch(
        &self,
        cb: impl FnOnce(&mut Self::Writer) + Send,
    ) -> impl Future<Output = ()> + Send;
    fn phase(&self) -> impl Future<Output = FlexiLogPhase> + Send;

    // Provided methods
    fn set_progress(&self, progress: f64) -> impl Future<Output = ()> + Send { ... }
    fn set_phase(&self, phase: FlexiLogPhase) -> impl Future<Output = ()> + Send { ... }
    fn log_with_opts(
        &self,
        lvl: Level,
        msg: String,
        force_replace_prior: bool,
    ) -> impl Future<Output = ()> + Send { ... }
    fn log_debug(
        &self,
        msg: impl Into<String>,
    ) -> impl Future<Output = ()> + Send { ... }
    fn log_info(
        &self,
        msg: impl Into<String>,
    ) -> impl Future<Output = ()> + Send { ... }
    fn log_warn(
        &self,
        msg: impl Into<String>,
    ) -> impl Future<Output = ()> + Send { ... }
    fn log_error(
        &self,
        msg: impl Into<String>,
    ) -> impl Future<Output = ()> + Send { ... }
    fn set_failed(&self) -> impl Future<Output = ()> + Send { ... }
    fn set_completed(&self) -> impl Future<Output = ()> + Send { ... }
    fn set_running(&self) -> impl Future<Output = ()> + Send { ... }
    fn set_failed_pending_retry(
        &self,
        scheduled_for: DateTime<Utc>,
    ) -> impl Future<Output = ()> + Send { ... }

}

An interface that can be used to track arbitrary logging “flows”.

Required Associated Types

type Writer: FlexiLogWriter

The actual object used to apply logs.

Required Methods

fn batch( &self, cb: impl FnOnce(&mut Self::Writer) + Send, ) -> impl Future<Output = ()> + Send

Can be used to synchronously batch together multiple changes.

fn phase(&self) -> impl Future<Output = FlexiLogPhase> + Send

Get the current phase. Just always return e,g, Pending if not supported.

Provided Methods

fn set_progress(&self, progress: f64) -> impl Future<Output = ()> + Send

Set the current progress. Depending on the underlying imp this could e.g. log a message, or write a progress bar etc.

fn set_phase(&self, phase: FlexiLogPhase) -> impl Future<Output = ()> + Send

Set the current phase. The underlying imp might not do anything for phases.

fn log_with_opts( &self, lvl: Level, msg: String, force_replace_prior: bool, ) -> impl Future<Output = ()> + Send

Underlying log adder with more options. If the actual output doesn’t support force_replace_prior, just ignore it and add like normal. (Same with lvl.)

fn log_debug(&self, msg: impl Into<String>) -> impl Future<Output = ()> + Send

Log a debug message.

fn log_info(&self, msg: impl Into<String>) -> impl Future<Output = ()> + Send

Log an info message.

fn log_warn(&self, msg: impl Into<String>) -> impl Future<Output = ()> + Send

Log a warning message.

fn log_error(&self, msg: impl Into<String>) -> impl Future<Output = ()> + Send

Log an error message.

fn set_failed(&self) -> impl Future<Output = ()> + Send

Set the state of the flow to failed.

fn set_completed(&self) -> impl Future<Output = ()> + Send

Set the state of the flow to completed.

fn set_running(&self) -> impl Future<Output = ()> + Send

Set the state of the flow to running.

fn set_failed_pending_retry( &self, scheduled_for: DateTime<Utc>, ) -> impl Future<Output = ()> + Send

Set the state of the flow to failed pending retry.

Object Safety

This trait is not object safe.