Trait bitbazaar::misc::FlexiLogFromRedis

pub trait FlexiLogFromRedis: Debug {
    type FlexiLogger: FlexiLog;

    // Required method
    fn into_flexi_log(self, redis: &Redis) -> Self::FlexiLogger;
}

Lots of loggers are backed by redis. A higher-order trait to produce a logger once redis is available.

Required Associated Types

type FlexiLogger: FlexiLog

The type of the flexi logger.

Required Methods

fn into_flexi_log(self, redis: &Redis) -> Self::FlexiLogger

Create a new flexi logger from a redis connection.