Trait bitbazaar::misc::FlexiLogWriter

pub trait FlexiLogWriter: Debug {
    // Required methods
    fn phase(&self) -> FlexiLogPhase;
    fn set_progress(&self, progress: f64);
    fn set_phase(&self, phase: FlexiLogPhase);
    fn log_with_opts(&self, lvl: Level, msg: String, force_replace_prior: bool);

    // Provided methods
    fn log_debug(&self, msg: impl Into<String>) { ... }
    fn log_info(&self, msg: impl Into<String>) { ... }
    fn log_warn(&self, msg: impl Into<String>) { ... }
    fn log_error(&self, msg: impl Into<String>) { ... }
    fn set_failed(&self) { ... }
    fn set_completed(&self) { ... }
    fn set_running(&self) { ... }
    fn set_failed_pending_retry(&self, scheduled_for: DateTime<Utc>) { ... }
}

The actual object used to apply logs. This is useful when updates are async, making synchronous batching easy.

Required Methods

fn phase(&self) -> FlexiLogPhase

Get the current phase. Just always return e,g, Pending if not supported.

fn set_progress(&self, progress: f64)

Set the current progress. Depending on the underlying imp this could e.g. log a message, or write a progress bar etc.

fn set_phase(&self, phase: FlexiLogPhase)

Set the current phase. The underlying imp might not do anything for phases.

fn log_with_opts(&self, lvl: Level, msg: String, force_replace_prior: bool)

Underlying log adder with more options. If the actual output doesn’t support force_replace_prior, just ignore it and add like normal. (Same with lvl.)

Provided Methods

fn log_debug(&self, msg: impl Into<String>)

Log a debug message.

fn log_info(&self, msg: impl Into<String>)

Log an info message.

fn log_warn(&self, msg: impl Into<String>)

Log a warning message.

fn log_error(&self, msg: impl Into<String>)

Log an error message.

fn set_failed(&self)

Set the state of the flow to failed.

fn set_completed(&self)

Set the state of the flow to completed.

fn set_running(&self)

Set the state of the flow to running.

fn set_failed_pending_retry(&self, scheduled_for: DateTime<Utc>)

Set the state of the flow to failed pending retry.

Object Safety

This trait is not object safe.