Trait bitbazaar::misc::IterWithCloneLazy

pub trait IterWithCloneLazy {
    type IterT;

    // Required method
    fn with_clone_lazy<ItemT: Clone>(
        self,
        item: ItemT,
    ) -> impl Iterator<Item = (Self::IterT, ItemT)>
       where Self: Sized;
}

Efficient way to clone an item for each element in an iterator. The final iteration will consume the original item, so no unnecessary clones are made.

Required Associated Types

type IterT

The return type of the iterator.

Required Methods

fn with_clone_lazy<ItemT: Clone>( self, item: ItemT, ) -> impl Iterator<Item = (Self::IterT, ItemT)>
where Self: Sized,

Efficient way to pass an owned clone of an item to each element in an iterator. Will pass the final item by value without cloning, so no unnecessary clones are made.