Module bitbazaar::redis

Redis utilities

Re-exports

pub use deadpool_redis;

pub use redis;

Structs

Redis

A wrapper around redis to make it more concise to use and not need redis in the downstream Cargo.toml.

RedisBatch

A command builder struct. Committed with RedisBatch::fire.

RedisChannelListener

A listener to receive messages from a redis channel via pubsub.

RedisConn

A lazy redis connection.

RedisJson

A wrapper on an arbitrary json object to allow reading and writing to redis. Access the inner with .0.

RedisJsonBorrowed

A borrowed wrapper on an arbitrary json object to writing to redis. Use this over RedisJson when you don’t want to own the data and prevent unnecessary cloning. Access the inner with .0.

RedisLock

A distributed lock for Redis.

RedisScript

A lua script wrapper. Should be created once per script.

RedisScriptInvoker

Represents a individual script call with specific args and keys.

RedisStandalone

Standalone redis client, using a unique free port. Useful for testing.

RedisTempList

Connect up to a magic redis list that:

RedisTempListItem

A user friendly interface around a redis list item, allowing for easy updates and replacements. This encapsulates when items aren’t available, or the user doesn’t actually want to use an item in some cases, but doesn’t want to pass Options<> around.

RedisTempListItemWithConn

A wrapped item, with a connection too, preventing need to pass 2 things around if useful for certain interfaces.

Enums

RedisLockErr

Errors that can occur when trying to lock a resource.

RedisTempListChangeEvent

Change events sent through a redis pubsub channel.

Traits

RedisBatchFire

Trait implementing the fire() method on a batch, variable over the items in the batch.

RedisBatchReturningOps

Implements all the supported redis operations that need to modify the return type and hence need macros.

RedisConnLike

Generic methods over the RedisConn and RedisConnOwned types.