Struct bitbazaar::redis::Redis

pub struct Redis { /* private fields */ }

A wrapper around redis to make it more concise to use and not need redis in the downstream Cargo.toml.

This wrapper attempts to return very few errors to help build in automatic redis failure handling into downstream code. All redis errors (availability, unexpected content) will be logged as errors and results returned as None (or similar) where possible.

Implementations

impl Redis

pub fn new( redis_conn_str: impl Into<String>, prefix: impl Into<String>, ) -> RResult<Self, AnyErr>

Create a new global redis wrapper from the given Redis URL (like redis://127.0.0.1).

Note this should only be done once at startup.

pub fn conn(&self) -> RedisConn<'_>

Get a RedisConn redis can be called with.

pub async fn dlock( &self, namespace: &'static str, lock_key: &str, time_to_live: TimeDelta, wait_up_to: Option<TimeDelta>, ) -> RResult<RedisLock<'_>, RedisLockErr>

Get a distributed redis lock.

This lock will prevent others getting the lock, until itโs time to live expires. Or the lock is manually released with RedisLock::unlock.

Arguments:

namespace: The redis key namespace to use.

lock_key: The resource to lock. Will be used as the key in Redis.

ttl: The time to live for this lock. After this time, the lock will be automatically released.

wait_up_to: if the lock is busy elsewhere, wait this long trying to get it, before giving up and returning RedisLockErr::Unavailable.

pub async fn dlock_for_fut<R, Fut: Future<Output = RResult<R, AnyErr>>>( &self, namespace: &str, lock_key: &str, wait_up_to: Option<TimeDelta>, fut: Fut, ) -> RResult<R, RedisLockErr>

Get a distributed redis lock that is held for the duration of the closure. The lock will be automatically released when the closure finishes.

Arguments:

namespace: The redis key namespace to use.

lock_key: The resource to lock. Will be used as the key in Redis.

wait_up_to: if the lock is busy elsewhere, wait this long trying to get it, before giving up and returning RedisLockErr::Unavailable.

pub fn get_inner_pool(&self) -> &Pool

Escape hatch, access the inner deadpool_redis pool.

pub fn get_conn_to_each_server(&self) -> Vec<RedisConn<'_>>

Used for dlock, the dlock algo is setup with multiple servers in mind, and synchronising locking between them. Itโs a good, future proofed algo, so keeping the multi interface despite the current implementation only using one server.