Struct bitbazaar::redis::RedisBatch

pub struct RedisBatch<'a, 'c, ConnType: RedisConnLike, ReturnType> { /* private fields */ }

A command builder struct. Committed with RedisBatch::fire.

Batched commands are run in order, but other commands from different sources may be interleaved. Note each command may be run twice, if scripts needed caching to redis.

Implementations

impl<'a, 'c, ConnType: RedisConnLike, ReturnType> RedisBatch<'a, 'c, ConnType, ReturnType>

pub fn script_no_return(self, script_invokation: RedisScriptInvoker<'c>) -> Self

Run an arbitrary redis (lua script). But discards any return value.

pub fn custom_no_return(self, cmd: &str) -> Self

Low-level backdoor. Pass in a custom redis command to run, but don’t expect a return value. After calling this, custom_arg() can be used to add arguments.

E.g. batch.custom_no_return("SET").custom_arg("key").custom_arg("value").fire().await;

pub fn custom_arg(self, arg: impl ToRedisArgs) -> Self

Low-level backdoor. Add a custom argument to the last custom command added with either custom_no_return() or custom().

pub fn publish( self, namespace: &str, channel: &str, message: impl Serialize + for<'b> Deserialize<'b>, ) -> Self

Publish a message to a pubsub channel. Json wrapped internally, makes easier to work with.

pub fn expire(self, namespace: &str, key: &str, ttl: Duration) -> Self

Expire an existing key with a new/updated ttl.

https://redis.io/commands/pexpire/

pub fn hset<S: AsRef<str>, Value: ToRedisArgs>( self, hashmap_namespace: &str, hashmap_key: &str, hashmap_ttl: Option<Duration>, entries: impl IntoIterator<Item = (S, Value)>, ) -> Self

Add entries to a hashmap (auto creating the hashmap if it doesn’t exist). https://redis.io/docs/latest/commands/hset/

Arguments:

hashmap_namespace: The namespace of the hashmap.

hashmap_key: The key of the hashmap.

hashmap_ttl: The time to live of the hashmap. This will reset on each addition, meaning after the last update the hashmap will expire after this time.

entries: The key/value pairs to add.

pub fn sadd<S: AsRef<str>>( self, set_namespace: &str, set_key: &str, set_ttl: Option<Duration>, items: impl IntoIterator<Item = S>, ) -> Self

Add an entry to a set (auto creating the set if it doesn’t exist). https://redis.io/commands/sadd/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

set_ttl: The time to live of the set. This will reset on each addition, meaning after the last update the set will expire after this time.

items: The items to add, values of sets must be strings.

pub fn srem<S: AsRef<str>>( self, set_namespace: &str, set_key: &str, items: impl IntoIterator<Item = S>, ) -> Self

remove an entries from a set. https://redis.io/commands/srem/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

values: The values to remove as an iterator.

pub fn zadd( self, set_namespace: &str, set_key: &str, set_ttl: Option<Duration>, score: i64, value: impl ToRedisArgs, ) -> Self

Add an entry to an ordered set (auto creating the set if it doesn’t exist). https://redis.io/commands/zadd/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

set_ttl: The time to live of the set. This will reset on each addition, meaning after the last update the set will expire after this time.

score: The score of the entry.

value: The value of the entry. (values of sets must be strings)

pub fn zrem<T: ToRedisArgs>( self, set_namespace: &str, set_key: &str, values: impl IntoIterator<Item = T>, ) -> Self

remove an entries from an ordered set. https://redis.io/commands/zrem/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

values: The values to remove as an iterator.

pub fn zadd_multi( self, set_namespace: &str, set_key: &str, set_ttl: Option<Duration>, items: &[(i64, impl ToRedisArgs)], ) -> Self

Add multiple entries at once to an ordered set (auto creating the set if it doesn’t exist). https://redis.io/commands/zadd/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

set_ttl: The time to live of the set. This will reset on each addition, meaning after the last update the set will expire after this time.

items: The scores and values of the entries. (set values must be strings)

pub fn zremrangebyscore( self, set_namespace: &str, set_key: &str, min: i64, max: i64, ) -> Self

Remove entries from an ordered set by score range. (range is inclusive)

https://redis.io/commands/zremrangebyscore/

pub fn set<T: ToRedisArgs>( self, namespace: &str, key: &str, value: T, expiry: Option<Duration>, ) -> Self

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<Value: ToRedisArgs>( self, namespace: &str, pairs: impl IntoIterator<Item = (impl AsRef<str>, Value)>, expiry: Option<Duration>, ) -> Self

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<S: AsRef<str>>( self, namespace: &str, keys: impl IntoIterator<Item = S>, ) -> Self

Clear one or more keys.

pub fn clear_namespace(self, namespace: &str) -> Self

Clear all keys under a given namespace