Trait bitbazaar::redis::RedisBatchFire

pub trait RedisBatchFire {
    type ReturnType;

    // Required method
    fn fire(self) -> impl Future<Output = Option<Self::ReturnType>>;
}

Available on crate feature redis only.

Trait implementing the fire() method on a batch, variable over the items in the batch.

Required Associated Types

type ReturnType

The final return type of the batch.

Required Methods

fn fire(self) -> impl Future<Output = Option<Self::ReturnType>>

Commit the batch and return the result. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

Object Safety

This trait is not object safe.