Trait bitbazaar::redis::RedisBatchOps

pub trait RedisBatchOps<'c> {
    type CurrentType;
    type NextType<T>;

    // Required methods
    fn script<ScriptOutput: FromRedisValue>(
        self,
        script_invokation: RedisScriptInvoker<'c>
    ) -> Self::NextType<ScriptOutput>;
    fn script_no_return(
        self,
        script_invokation: RedisScriptInvoker<'c>
    ) -> Self::CurrentType;
    fn set(
        self,
        namespace: &'static str,
        key: &str,
        value: impl ToRedisArgs,
        expiry: Option<Duration>
    ) -> Self::CurrentType;
    fn mset<'key, Value: ToRedisArgs>(
        self,
        namespace: &'static str,
        pairs: impl IntoIterator<Item = (&'key str, Value)>,
        expiry: Option<Duration>
    ) -> Self::CurrentType;
    fn clear<'key>(
        self,
        namespace: &'static str,
        keys: impl IntoIterator<Item = &'key str>
    ) -> Self::CurrentType;
    fn clear_namespace(self, namespace: &'static str) -> Self::CurrentType;
    fn exists(self, namespace: &'static str, key: &str) -> Self::NextType<bool>;
    fn mexists<'key>(
        self,
        namespace: &'static str,
        keys: impl IntoIterator<Item = &'key str>
    ) -> Self::NextType<Vec<bool>>;
    fn get<Value: FromRedisValue>(
        self,
        namespace: &'static str,
        key: &str
    ) -> Self::NextType<Option<Value>>;
    fn mget<'key, Value>(
        self,
        namespace: &'static str,
        keys: impl IntoIterator<Item = &'key str>
    ) -> Self::NextType<Vec<Option<Value>>>;
}

Available on crate feature redis only.

Implements all the supported redis operations for the batch.

Required Associated Types

type CurrentType

The current batch struct sig.

type NextType<T>

The producer for the next batch struct sig.

Required Methods

fn script<ScriptOutput: FromRedisValue>( self, script_invokation: RedisScriptInvoker<'c> ) -> Self::NextType<ScriptOutput>

Run an arbitrary redis (lua script).

fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> Self::CurrentType

Run an arbitrary redis (lua script). But discards any return value.

fn set( self, namespace: &'static str, key: &str, value: impl ToRedisArgs, expiry: Option<Duration> ) -> Self::CurrentType

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

fn mset<'key, Value: ToRedisArgs>( self, namespace: &'static str, pairs: impl IntoIterator<Item = (&'key str, Value)>, expiry: Option<Duration> ) -> Self::CurrentType

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

fn clear<'key>( self, namespace: &'static str, keys: impl IntoIterator<Item = &'key str> ) -> Self::CurrentType

Clear one or more keys.

fn clear_namespace(self, namespace: &'static str) -> Self::CurrentType

Clear all keys under a given namespace

fn exists(self, namespace: &'static str, key: &str) -> Self::NextType<bool>

Check if a key exists.

fn mexists<'key>( self, namespace: &'static str, keys: impl IntoIterator<Item = &'key str> ) -> Self::NextType<Vec<bool>>

Check if multiple keys exists.

fn get<Value: FromRedisValue>( self, namespace: &'static str, key: &str ) -> Self::NextType<Option<Value>>

Get a value from a key. Returning None if the key doesn’t exist.

fn mget<'key, Value>( self, namespace: &'static str, keys: impl IntoIterator<Item = &'key str> ) -> Self::NextType<Vec<Option<Value>>>

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

Object Safety

This trait is not object safe.