Trait bitbazaar::redis::RedisConnLike

pub trait RedisConnLike: Debug + Send + Sized {
    // Required methods
    async fn get_inner_conn(&mut self) -> Option<&mut Connection>;
    fn prefix(&self) -> &str;
    fn into_owned(self) -> RedisConnOwned;

    // Provided methods
    async fn ping(&mut self) -> bool { ... }
    async fn dev_flushall(&mut self, sync: bool) -> Option<String> { ... }
    async fn rate_limiter(
        &mut self,
        namespace: &str,
        caller_identifier: &str,
        start_delaying_after_attempt: usize,
        initial_delay: Duration,
        multiplier: f64,
    ) -> Option<Duration> { ... }
    fn batch(&mut self) -> RedisBatch<'_, '_, Self, ()> { ... }
    fn final_namespace(&self, namespace: &str) -> String { ... }
    fn final_key(&self, namespace: &str, key: Cow<'_, str>) -> String { ... }
    async fn cached_fn<'b, T, Fut, K: Into<Cow<'b, str>>>(
        &mut self,
        namespace: &str,
        key: K,
        expiry: Option<Duration>,
        cb: impl FnOnce() -> Fut,
    ) -> RResult<T, AnyErr>
       where T: FromRedisValue + ToRedisArgs,
             Fut: Future<Output = Result<T, AnyErr>> { ... }
}

Generic methods over the RedisConn and RedisConnOwned types.

Required Methods

async fn get_inner_conn(&mut self) -> Option<&mut Connection>

Get an internal connection from the pool, connections are kept in the pool for reuse. If redis is acting up and unavailable, this will return None. NOTE: this mainly is used internally, but provides a fallback to the underlying connection, if the exposed interface does not provide options that fit an external user need (which could definitely happen).

fn prefix(&self) -> &str

Get the redis configured prefix.

fn into_owned(self) -> RedisConnOwned

Convert to the owned variant.

Provided Methods

async fn ping(&mut self) -> bool

Ping redis, returning true if it’s up.

async fn dev_flushall(&mut self, sync: bool) -> Option<String>

Flush the whole redis cache, will delete all data.

async fn rate_limiter( &mut self, namespace: &str, caller_identifier: &str, start_delaying_after_attempt: usize, initial_delay: Duration, multiplier: f64, ) -> Option<Duration>

A simple rate_limiter/backoff helper. Can be used to protect against repeated attempts in quick succession. Once start_delaying_after_attempt is hit, the operation delay will multiplied by the multiplier each time. Only once no call is made for the duration of the current delay (so current delay doubled) will the attempt number reset to zero.

Arguments:

namespace: A unique identifier for the endpoint, e.g. user-login.

caller_identifier: A unique identifier for the caller, e.g. a user id.

start_delaying_after_attempt: The number of attempts before the delays start being imposed.

initial_delay: The initial delay to impose.

multiplier: The multiplier to apply, (attempt-start_delaying_after_attempt) * multiplier * initial_delay = delay.

Returns:

None: Continue with the operation.

Some<chrono::Duration>: Retry after the duration.

fn batch(&mut self) -> RedisBatch<'_, '_, Self, ()>

Get a new RedisBatch for this connection that commands can be piped together with.

fn final_namespace(&self, namespace: &str) -> String

Redis keys are all prefixed, use this to finalise a namespace outside of built in commands, e.g. for use in a custom script.

fn final_key(&self, namespace: &str, key: Cow<'_, str>) -> String

Redis keys are all prefixed, use this to finalise a key outside of built in commands, e.g. for use in a custom script.

async fn cached_fn<'b, T, Fut, K: Into<Cow<'b, str>>>( &mut self, namespace: &str, key: K, expiry: Option<Duration>, cb: impl FnOnce() -> Fut, ) -> RResult<T, AnyErr>
where T: FromRedisValue + ToRedisArgs, Fut: Future<Output = Result<T, AnyErr>>,

Cache an async function in redis with an optional expiry. If already stored, the cached value will be returned, otherwise the function will be stored in redis for next time.

If redis is unavailable, or the existing contents at the key is wrong, the function output will be used. The only error coming out of here should be something wrong with the external callback.

Expiry accurate to a millisecond.

Object Safety

This trait is not object safe.