Trait bitbazaar::redis::RedisJsonConsume

pub trait RedisJsonConsume<T> {
    // Required method
    fn consume(self) -> T;
}

Available on crate feature redis only.

A trait to allow consuming a redis wrapper and returning the inner json object.

Required Methods

fn consume(self) -> T

Extract the inner json object from the redis wrapper.