Crate bitbazaar

BitBazaar - A crate containing miscellaneous public utilities.

Modules

cli

Command line interface utilities.

errors

Error handling utilities.

hash

Hashing utilities.

logging

Logging utilities

timing

Timing utilities

Macros

err

A macro for creating a TracedErr from a string or another existing error.

panic_on_err

When working in a function that cannot return a result, wrap a block in this macro to panic with the formatted error if it errors.

spez

Spez is a macro used inside our err! macro, need to export from the bb crate itself to prevent downstream needing to import it: Specialize based on the type of an expression.

timeit

Time a block of code and log to the global time recorder.

warn_user_once

Warn a user once, with uniqueness determined by the calling location itself.

warn_user_once_by_id

Warn a user once, with uniqueness determined by the given ID.