Crate bitbazaar

bitbazaar - An assortment of publicly available cross-language utilities useful to my projects.

Modules

clicli

Command line interface utilities.

errors

Error handling utilities.

hash

Hashing utilities.

log

Logging utilities

misc

Completely miscellaneous utilities

redisredis

Redis utilities

timingtiming

Timing utilities

Macros

anyerr

A macro for building Report<AnyErr> objects with string information easily.

err

A macro for building Report<ArbitraryErrorStackErr> objects with string context easily.

panic_on_err

When working in a function that cannot return a result, use this to auto panic with the formatted error if something goes wrong.

panic_on_err_async

When working in a function that cannot return a result, use this to auto panic with the formatted error if something goes wrong.

timeittiming

Time a block of code and log to the global time recorder.

warn_user_once

Warn a user once, with uniqueness determined by the calling location itself.

warn_user_once_by_id

Warn a user once, with uniqueness determined by the given ID.