Module bitbazaar::misc

Completely miscellaneous utilities

Modules

bytes

Byte manipulation utilities, e.g. transfer speed.

Macros

retry_flexi

Internal version of Retry::call that:

Structs

PeriodicUpdater

A periodic updater. Run a callback at a specified time interval. Synchronous. Requires polling. Useful in long running loops to do something at a specified time interval. PeriodicUpdater::maybe_update should be called during each loop.

Refreshable

A data wrapper that automatically updates the data given out when deemed stale. The data is set to refresh at a certain interval (triggered on access), or can be forcefully refreshed.

RefreshableGuard

A temporary storage of the pointer.

Retry

The retry builder.

RetryInfo

Information about the last retry attempt.

Enums

Backoff

UNSTABLE: ONLY PUBLIC FOR MACRO USE.

RetryUntil

UNSTABLE: ONLY PUBLIC FOR MACRO USE.

Traits

BinarySearchable

A trait allowing arbitrary types to be binary searchable.

FlexiLog

An interface that can be used to track arbitrary logging “flows”.

SerdeMigratable

A trait to help with migrating data structures when they change.

Functions

binary_search_exact

Binary search for the desired item in the array. Returns None when not exactly found.

binary_search_exact_with_siblings

Binary search for the desired item in the array, but also passes the prior and next items into the comparison function, if they exist. Returns None when not exactly found.

binary_search_soft

Binary search for the desired item in the array. Only returns None when the array is empty.

in_ci

Returns true if the current process seems to be running in CI.

is_tcp_port_listening

Check if a port is listening for a given ipv4 address and port.

main_wrapper

Wrap your main function in this to auto handle common program needs.

sleep_compat

Sleep for a duration, compatible with both WASM and native targets.