Module bitbazaar::misc

Completely miscellaneous utilities

Modules

bytes

Byte manipulation utilities, e.g. transfer speed.

Structs

PeriodicUpdater

A periodic updater. Run a callback at a specified time interval. Synchronous. Requires polling. Useful in long running loops to do something at a specified time interval. PeriodicUpdater::maybe_update should be called during each loop.

RetryBackoffInfo

Information about the last retry attempt.

Traits

FlexiLog

An interface that can be used to track arbitrary logging “flows”.

Functions

binary_search_exact

Binary search for the desired item in the array. Returns None when not exactly found.

binary_search_soft

Binary search for the desired item in the array. Only returns None when the array is empty.

in_ci

Returns true if the current process seems to be running in CI.

is_tcp_port_listening

Check if a port is listening for a given ipv4 address and port.

retry_backoff

Will attempt execute the fn’s returned future according to the entered spec.

sleep_compat

Sleep for a duration, compatible with both WASM and native targets.