Struct bitbazaar::redis::RedisBatch

pub struct RedisBatch<'a, 'b, 'c, ReturnType> { /* private fields */ }

A command builder struct. Committed with RedisBatch::fire.

Batched commands are run in order, but other commands from different sources may be interleaved. Note each command may be run twice, if scripts needed caching to redis.

Implementations

impl<'a, 'b, 'c, ReturnType> RedisBatch<'a, 'b, 'c, ReturnType>

pub fn script_no_return(self, script_invokation: RedisScriptInvoker<'c>) -> Self

Run an arbitrary redis (lua script). But discards any return value.

pub fn expire(self, namespace: &str, key: &str, ttl: Duration) -> Self

Expire an existing key with a new/updated ttl.

https://redis.io/commands/pexpire/

pub fn zadd( self, set_namespace: &str, set_key: &str, set_ttl: Option<Duration>, score: i64, value: impl ToRedisArgs ) -> Self

Add an entry to an ordered set (auto creating the set if it doesn’t exist). https://redis.io/commands/zadd/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

set_ttl: The time to live of the set. This will reset on each addition, meaning after the last update the set will expire after this time.

score: The score of the entry.

value: The value of the entry. (values of sets must be strings)

pub fn zrem<T: ToRedisArgs>( self, set_namespace: &str, set_key: &str, values: impl IntoIterator<Item = T> ) -> Self

remove an entries from an ordered set. https://redis.io/commands/zrem/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

values: The values to remove as an iterator.

pub fn zadd_multi( self, set_namespace: &str, set_key: &str, set_ttl: Option<Duration>, items: &[(i64, impl ToRedisArgs)] ) -> Self

Add multiple entries at once to an ordered set (auto creating the set if it doesn’t exist). https://redis.io/commands/zadd/

Arguments:

set_namespace: The namespace of the set.

set_key: The key of the set.

set_ttl: The time to live of the set. This will reset on each addition, meaning after the last update the set will expire after this time.

items: The scores and values of the entries. (set values must be strings)

pub fn zremrangebyscore( self, set_namespace: &str, set_key: &str, min: i64, max: i64 ) -> Self

Remove entries from an ordered set by score range. (range is inclusive)

https://redis.io/commands/zremrangebyscore/

pub fn set<T: ToRedisArgs>( self, namespace: &str, key: &str, value: T, expiry: Option<Duration> ) -> Self

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<Value: ToRedisArgs>( self, namespace: &str, pairs: impl IntoIterator<Item = (impl AsRef<str>, Value)>, expiry: Option<Duration> ) -> Self

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key>( self, namespace: &str, keys: impl IntoIterator<Item = &'key str> ) -> Self

Clear one or more keys.

pub fn clear_namespace(self, namespace: &str) -> Self

Clear all keys under a given namespace