Struct bitbazaar::redis::RedisBatch

pub struct RedisBatch<'a, 'b, 'c, ReturnType = ()> { /* private fields */ }

Available on crate feature redis only.

A command builder struct. Committed with RedisBatch::fire.

Batched commands are run in order, but other commands from different sources may be interleaved. Note each command may be run twice, if scripts needed caching to redis.

Implementations

impl<'a, 'b, 'c, R: FromRedisValue> RedisBatch<'a, 'b, 'c, (R,)>

pub async fn fire(self) -> Option<R>
where (R,): FromRedisValue,

Commit the batch and return the result. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

Note this is the special singular variant that returns the command output directly (no tuple).

impl<'a, 'b, 'c> RedisBatch<'a, 'b, 'c, ()>

pub async fn fire(self) -> Option<()>
where (): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B)>

pub async fn fire(self) -> Option<(A, B)>
where (A, B): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C)>

pub async fn fire(self) -> Option<(A, B, C)>
where (A, B, C): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D)>

pub async fn fire(self) -> Option<(A, B, C, D)>
where (A, B, C, D): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>

pub async fn fire(self) -> Option<(A, B, C, D, E)>
where (A, B, C, D, E): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F)>
where (A, B, C, D, E, F): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G)>
where (A, B, C, D, E, F, G): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H)>
where (A, B, C, D, E, F, G, H): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I)>
where (A, B, C, D, E, F, G, H, I): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I, J)>
where (A, B, C, D, E, F, G, H, I, J): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I, J, K)>
where (A, B, C, D, E, F, G, H, I, J, K): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I, J, K, L)>
where (A, B, C, D, E, F, G, H, I, J, K, L): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I, J, K, L, M)>
where (A, B, C, D, E, F, G, H, I, J, K, L, M): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue, N: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I, J, K, L, M, N)>
where (A, B, C, D, E, F, G, H, I, J, K, L, M, N): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue, N: FromRedisValue, O: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>

pub async fn fire(self) -> Option<(A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>
where (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue, N: FromRedisValue, O: FromRedisValue, P: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>

pub async fn fire( self ) -> Option<(A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>
where (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P): FromRedisValue,

Commit the batch and return the command results in a tuple. If redis unavailable, or the types didn’t match causing decoding to fail, None will be returned and the error logged.

impl<'a, 'b, 'c> RedisBatch<'a, 'b, 'c, ()>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (ScriptOutput,)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, ()>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, ()>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, ()>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, ()>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, ()>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (bool,)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (Vec<bool>,)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (Option<Value>,)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (Vec<Option<Value>>,)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue> RedisBatch<'a, 'b, 'c, (A,)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A,)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A,)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A,)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A,)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A,)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue, N: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue, N: FromRedisValue, O: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.

impl<'a, 'b, 'c, A: FromRedisValue, B: FromRedisValue, C: FromRedisValue, D: FromRedisValue, E: FromRedisValue, F: FromRedisValue, G: FromRedisValue, H: FromRedisValue, I: FromRedisValue, J: FromRedisValue, K: FromRedisValue, L: FromRedisValue, M: FromRedisValue, N: FromRedisValue, O: FromRedisValue, P: FromRedisValue> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>

pub fn script<ScriptOutput>( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, ScriptOutput)>
where ScriptOutput: FromRedisValue,

Run an arbitrary redis (lua script).

pub fn script_no_return( self, script_invokation: RedisScriptInvoker<'c> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>

Run an arbitrary redis (lua script). But discards any return value.

pub fn set<'key, Key, Value>( self, namespace: &'static str, key: Key, value: Value, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>
where Key: Into<Cow<'key, str>>, Value: ToRedisArgs,

Set a key to a value with an optional expiry.

(expiry accurate to the millisecond)

pub fn mset<'key, Key, Value, Pairs>( self, namespace: &'static str, pairs: Pairs, expiry: Option<Duration> ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>
where Value: ToRedisArgs, Key: Into<Cow<'key, str>>, Pairs: IntoIterator<Item = (Key, Value)>,

Set multiple values (MSET) of the same type at once. If expiry used will use a custom lua script to achieve the functionality.

(expiry accurate to the millisecond)

pub fn clear<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Clear one or more keys.

pub fn clear_namespace( self, namespace: &'static str ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P)>

Clear all keys under a given namespace

pub fn exists<'key, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, bool)>
where Key: Into<Cow<'key, str>>,

Check if a key exists.

pub fn mexists<'key, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Vec<bool>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>,

Check if multiple keys exists.

pub fn get<'key, Value, Key>( self, namespace: &'static str, key: Key ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Option<Value>)>
where Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get a value from a key. Returning None if the key doesn’t exist.

pub fn mget<'key, Value, Keys, Key>( self, namespace: &'static str, keys: Keys ) -> RedisBatch<'a, 'b, 'c, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Vec<Option<Value>>)>
where Keys: IntoIterator<Item = Key>, Key: Into<Cow<'key, str>>, Value: FromRedisValue,

Get multiple values (MGET) of the same type at once. Returning None for each key that didn’t exist.