Module bitbazaar::threads

Concurrency/parallelism utilities

Enums

BatchLimit

How batched futures are limited, either using a parent’s limit to prevent concurrency explosion, or direct.

Functions

batch_futures_descendants

Batch run futures but with a limiter on parents and descendants. IF HIGHEST PERFORMANCE IS NEEDED AND YOU DON’T NEED DESCENDANT LIMITING, USE [batch_futures_stream_async_cb_perf] or just normal [futures::StreamExt::buffer_unordered].

batch_futures_descendants_stream_async_cb

Underlying of batch_futures_descendants, use if need to process in order during execution.

batch_futures_descendants_stream_sync_cb

Underlying of batch_futures_descendants, use if need to process in order during execution.

batch_futures_flat

A simple and performant runner for an iterator of future creators. The futures are run in parallel up to the supplied limit. The results are returned in the same order as the inputted futures.

batch_futures_flat_stream_async_cb

A more performant version of batch_futures_descendants due to no descendant requirement. By removing descendant limiting, semaphores no longer need to be used.

batch_futures_flat_stream_sync_cb

A more performant version of batch_futures_descendants due to no descendant requirement. By removing descendant limiting, semaphores no longer need to be used.

run_cpu_intensive

The following link explains the distinction between normal tokio, [tokio::task::spawn_blocking], and rayon. For most cases you think to use [tokio::task::spawn_blocking], but isn’t actual IO like file read/write that isn’t async, and actually CPU bound stuff, rayon should be used. It goes without saying, if you ever want to use anything from rayon like [rayon::iter::ParallelIterator], it should be inside run_cpu_intensive https://ryhl.io/blog/async-what-is-blocking/#the-rayon-crate